added more type to the client sdk

+++ src/framework/index.ts
@@ -8,14 +8,12 @@ import type {
   DataQueryParams,
   FrameworkConfig,
   FrameworkEndpoints,
-  ModifySchemaParams,
   PermissionParams,
   SchemaCreateParams,
   TablePermissions,
 } from "./types.js";
 import type { UserContext } from "./types.js";
 import { createColumn } from "./database/column-utils.js";
-import { modifySchema } from "./database/schema.js";
 import { QueryHandler } from "./database/sdk/server.js";
 
 export class Framework {
@@ -105,12 +103,13 @@ export class Framework {
       },
       create: async (payload: SchemaCreateParams) => {
         try {
-          const { tableName, columns } = payload;
+          const { action, tableName, columns } = payload;
 
-          if (!tableName) {
+          if (!action || !tableName) {
             throw new Error("Invalid request body");
           }
 
+          if (action === "create") {
             await this.hooks
               .getKnexInstance()
               .schema.createTableIfNotExists(tableName, (table) => {
@@ -124,38 +123,25 @@ export class Framework {
             return {
               message: "Table created successfully",
               tablename: tableName,
-              action: 'create',
+              action,
             };
-
-        } catch (error: any) {
-          throw new Error(error);
-        }
-      },
-      delete: async (tableName: string) => {
-        try {
-          await this.hooks
+          } else if (action === "delete") {
+            await this.hooks
               .getKnexInstance()
               .schema.dropTableIfExists(tableName);
-
-            await this.permissionService.deletePermissionsForTable(tableName)
 
             return {
               message: "Table deleted successfully",
               tablename: tableName,
-              action: 'delete',
+              action,
             };
-        } catch(error: any) {
-           throw new Error(error)
-        }
-      },
-      modify: async (payload: ModifySchemaParams) => {
-        try {          
-          return await modifySchema(this.hooks.getKnexInstance(), payload);
+          }
+
+          throw new Error("Invalid action");
         } catch (error: any) {
           throw new Error(error);
         }
       },
-
     },
 
     data: {
